+++ src/main/java/fr/utbm/fisa/communicationapi/infrastructure/repositories/CommunicationRepository.java
@@ -3,5 +3,5 @@ package fr.utbm.fisa.communicationapi.infrastructure.repositories;
 import fr.utbm.fisa.communicationapi.infrastructure.entities.Communication;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CommunicationRepository extends CrudRepository<Communication, Integer> {
+public interface CommunicationRepository extends CrudRepository<Communication, Long> {
 }
