+++ src/main/java/fr/utbm/fisa/communicationapi/infrastructure/repositories/MessageRepository.java
@@ -2,13 +2,11 @@ package fr.utbm.fisa.communicationapi.infrastructure.repositories;
 
 import fr.utbm.fisa.communicationapi.infrastructure.entities.Discussion;
 import fr.utbm.fisa.communicationapi.infrastructure.entities.Message;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface MessageRepository extends CrudRepository<Message, Integer> {
+public interface MessageRepository extends CrudRepository<Message, Long> {
     Message findMessageByIdDiscussion(Discussion discussion);
 
     List<Message> findMessagesByIdDiscussion(Discussion discussion);
