+++ src/main/java/fr/utbm/fisa/communicationapi/infrastructure/repositories/PollAnswerRepository.java
@@ -2,12 +2,10 @@ package fr.utbm.fisa.communicationapi.infrastructure.repositories;
 
 import fr.utbm.fisa.communicationapi.infrastructure.entities.Poll;
 import fr.utbm.fisa.communicationapi.infrastructure.entities.PollAnswer;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface PollAnswerRepository extends CrudRepository<PollAnswer, Integer> {
+public interface PollAnswerRepository extends CrudRepository<PollAnswer, Long> {
     List<PollAnswer> findAllByIdPoll(Poll poll);
 }
